extract drag scroll component
