+++ Composer/packages/extensions/visual-designer/src/editors/ObiEditor.js
@@ -1,31 +1,16 @@
-import React, { useRef } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import { NodeEventTypes } from '../shared/NodeEventTypes';
 import { ObiTypes } from '../shared/ObiTypes';
 import { deleteNode, insert } from '../shared/jsonTracker';
+import DragScroll from '../components/DragScroll';
 
 import { AdaptiveDialogEditor } from './AdaptiveDialogEditor';
 import { RuleEditor } from './RuleEditor';
 import './ObiEditor.css';
 
-function getScrollParent(node) {
-  if (!node || node === document) return null;
-
-  const isNodeOverflow = node.scrollHeight > node.clientHeight || node.scrollWidth > node.clientWidth;
-  // REF: https://github.com/jquery/jquery-ui/blob/master/ui/scroll-parent.js
-  const isNodeScrollable = getComputedStyle(node).overflow.match(/(auto|scroll|hidden)/);
-
-  if (isNodeOverflow && isNodeScrollable) {
-    return node;
-  }
-  return getScrollParent(node.parentNode);
-}
-
 export const ObiEditor = ({ path, focusedId, data, onSelect, onExpand, onOpen, onChange }) => {
-  const isDragging = useRef(false);
-  const dragStartPoint = useRef({ x: 0, y: 0 });
-
   const dispatchEvent = (eventName, eventData) => {
     let handler;
     switch (eventName) {
@@ -88,45 +73,16 @@ export const ObiEditor = ({ path, focusedId, data, onSelect, onExpand, onOpen, o
         e.stopPropagation();
         dispatchEvent(NodeEventTypes.Focus, '');
       }}
-      onMouseDown={e => {
-        e.preventDefault();
-        e.stopPropagation();
-        isDragging.current = true;
-        dragStartPoint.current.x = e.pageX;
-        dragStartPoint.current.y = e.pageY;
-      }}
-      onMouseUp={e => {
-        e.preventDefault();
-        e.stopPropagation();
-        isDragging.current = false;
-      }}
-      onMouseLeave={() => {
-        isDragging.current = false;
-      }}
-      onMouseMove={e => {
-        if (isDragging.current) {
-          const mouseMove = {
-            x: e.pageX - dragStartPoint.current.x,
-            y: e.pageY - dragStartPoint.current.y,
-          };
-          const scrollParent = getScrollParent(e.currentTarget);
-          if (scrollParent) {
-            scrollParent.scrollBy(-mouseMove.x, -mouseMove.y);
-            dragStartPoint.current.x = e.pageX;
-            dragStartPoint.current.y = e.pageY;
-          } else {
-            window.scrollBy(-mouseMove.x, -mouseMove.y);
-          }
-        }
-      }}
     >
-      <ChosenEditor
-        id={path}
-        data={data}
-        expanded={true}
-        focusedId={focusedId}
-        onEvent={(...args) => dispatchEvent(...args)}
-      />
+      <DragScroll>
+        <ChosenEditor
+          id={path}
+          data={data}
+          expanded={true}
+          focusedId={focusedId}
+          onEvent={(...args) => dispatchEvent(...args)}
+        />
+      </DragScroll>
     </div>
   );
 };
